requirements.txt
discord.py
pathvalidate
mutagen
youtube_dl
aiogoogle
pynacl
loguru
Pillow
validator-collection
